+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixes to asyncio usage for Python 3.14 (Python 3.8 or higher), while maintaining backwards compatibility.
 - Fix relative import in `__main__.py`
+- Add URLs and request-method information to the final summary
 - Add missing requirements to `requirements.txt`
 - Fix flimsy packaging entrypoint
 - Misc: Create changelog